abstracting away things
